refactor: Move trips in different directory
