+++ src/components/TripsList.jsx
@@ -11,7 +11,7 @@ import { isQueryLoading, useQuery, hasQueryBeenLoaded } from 'cozy-client'
 
 import { buildGeoJSONQuery } from 'src/queries/queries'
 import TripItem from 'src/components/TripItem'
-import { transformTimeSeriesToTrips, getStartDate } from './trips'
+import { transformTimeSeriesToTrips, getStartDate } from 'src/lib/trips'
 
 export const TripsList = ({ account }) => {
   const { t } = useI18n()
