+++ src/lib/trips.js
