+++ src/lib/trips.spec.js
